+++ js/app-meta.js
@@ -1,6 +1,6 @@
 export const APP_META = Object.freeze({
   name: 'Nederlands, gewoon doen',
-  version: '19.0.1',
+  version: '19.0.2',
   storageSchema: 1,
-  cacheName: 'nederlands-gewoon-doen-v19-0-1'
+  cacheName: 'nederlands-gewoon-doen-v19-0-2'
 });